Move deprecated controllers. OrbitState refresh
