+++ src/core/deprecated/controllers/map-controls.js
