+++ src/core/deprecated/controllers/view-state.js
