+++ src/react/index.js
@@ -20,10 +20,9 @@
 
 export {default as DeckGL} from './deckgl';
 export {default} from './deckgl';
-
-// TODO - should react controllers be exported or just integrated into deck.gl API?
-export {default as ViewportController} from './viewport-controller';
-export {default as MapController} from './map-controller';
-export {default as OrbitController} from './experimental/orbit-controller';
-
 export {default as autobind} from './utils/autobind';
+
+// TODO - React controllers should just be integrated into deck.gl API
+export {default as MapController} from './deprecated/map-controller';
+export {default as OrbitController} from './deprecated/orbit-controller';
+export {default as ViewportController} from './deprecated/viewport-controller';
